somemore refactor
