+++ android/app/src/main/java/com/hedvig/app/feature/chat/ChatRepository.kt
@@ -1,12 +1,12 @@
-package com.hedvig.app.react.chat
+package com.hedvig.app.feature.chat
 
 import android.content.Context
 import android.net.Uri
 import com.google.gson.GsonBuilder
 import com.hedvig.app.service.FileService
 import com.hedvig.app.util.extensions.into
-import com.hedvig.app.react.chat.dto.UploadData
-import com.hedvig.app.react.chat.dto.UploadResponse
+import com.hedvig.app.feature.chat.dto.UploadData
+import com.hedvig.app.feature.chat.dto.UploadResponse
 import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 import okhttp3.MediaType
