+++ android/app/src/main/java/com/hedvig/app/feature/chat/dto/UploadError.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.react.chat.dto
+package com.hedvig.app.feature.chat.dto
 
 data class UploadError(
     val message: String?
