+++ android/app/src/main/java/com/hedvig/app/react/ActivityStarterReactPackage.kt
@@ -1,10 +1,11 @@
-package com.hedvig.app.starter
+package com.hedvig.app.react
 
 import com.apollographql.apollo.ApolloClient
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
+import com.hedvig.app.react.ActivityStarterModule
 import java.util.ArrayList
 
 class ActivityStarterReactPackage(private val apolloClient: ApolloClient) : ReactPackage {
