+++ android/app/src/main/java/com/hedvig/app/ui/fragment/RoundedBottomSheetDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.ui.common
+package com.hedvig.app.ui.fragment
 
 import android.app.Dialog
 import android.os.Bundle
