+++ android/app/src/main/java/com/hedvig/app/util/AppContextExtension.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.utils
+package com.hedvig.app.util
 
 import android.app.Dialog
 import android.content.Context
