+++ android/app/src/main/res/layout/cashback_option.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.hedvig.app.ui.common.HedvigCardView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.hedvig.app.ui.view.HedvigCardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -56,4 +56,4 @@
 
     </android.support.constraint.ConstraintLayout>
 
-</com.hedvig.app.ui.common.HedvigCardView>
+</com.hedvig.app.ui.view.HedvigCardView>
